RagTransformer/requirements.txt
transformers
torch
serpapi
python-dotenv